211 factory pattern for loaders (#220)
* update appveyor receipt - use conda env create -f dev_environment.yml - now try several python versions

* nothing

* move splitting of prms to own method in set_instrument

* 1) Allow for only one instrument id (for example only "arbin_res", not both "arbin" and "arbin_res")
2) Implement an instrument loader factory that now searches for the hard-coded instrument loaders and can create them (not moved to set_instrument yet)

* Factory implemented - but still a couple of bugs

* Factory implemented - but still a couple of bugs

* Factory implemented - bug fixed

* Changing naming scheme: all loader classes must from now on be called DataLoader

* Factory working

* Factory with doc strings

* Factory with doc strings
