+++ cellpy/parameters/.cellpy_prms_default.conf
@@ -139,7 +139,7 @@ Reader:
   use_cellpy_stat_file: false
   auto_dirs: true
 Instruments:
-  tester: arbin
+  tester: arbin_res
   custom_instrument_definitions_file:
   Arbin:
     chunk_size:
