+++ tests/test_sql.py
@@ -12,13 +12,13 @@ log.setup_logging(default_level="DEBUG", testing=True)
 @pytest.fixture
 def raw_mock(parameters):
     raw = pd.read_excel(parameters.mock_file_path, sheet_name="arbin_sql")
-    raw.Date_Time = raw.Date_Time.astype('Int64')
+    raw.Date_Time = raw.Date_Time.astype("Int64")
     return raw
 
 
 def test_import_arbin_sql():
-    loader = arbin_sql.ArbinSQLLoader()
-    assert isinstance(loader, base.Loader)
+    loader = arbin_sql.DataLoader()
+    assert isinstance(loader, base.BaseLoader)
 
 
 def test_post_process_rename_headers_defined():
@@ -28,7 +28,7 @@ def test_post_process_rename_headers_defined():
         "rename_headers": True,
         "extract_start_datetime": False,
     }
-    loader = arbin_sql.ArbinSQLLoader()
+    loader = arbin_sql.DataLoader()
 
     # creating a mock raw data
     data = Cell()
@@ -45,9 +45,8 @@ def test_post_process_rename_headers_from_file(raw_mock):
         "rename_headers": True,
         "extract_start_datetime": True,
     }
-    loader = arbin_sql.ArbinSQLLoader()
+    loader = arbin_sql.DataLoader()
 
     data = Cell()
     data.raw = raw_mock
     loader._post_process(data, **keywords)
-
